+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+- Added the `Calendar` component.
 
 ### Fixed
 - Fixed tooltip cropping in ProgressBar when asset names exceed container width.